pyproject.toml
[project]
name = "kanidm-discord-bot"
version = "2026.03.1"
description = "Discord bot for Kanidm user registration"
requires-python = ">=3.11"
dependencies = [
    "discord.py~=2.7",
    "httpx~=0.28",
    "python-dotenv~=1.2",
]
